(Hopefully) Proper audio filtering system
Samples are requested to the filter chain which then propagates the
requests to the needed filters. There's no high-level interface yet and
the test code uses internal stuff.

Right now only has a source filter, that is capable of returning almost
arbitrary samples from the input file. It doesn't handle backward seeking
requests yet, in case the requested samples are no longer in the cache.

It is not yet integrated with x264 but will be once encoding works.

Due to this architecture, x264's lavf demuxer will have to implement its
own audio decoding.