Probably Overcomplicated high-level-ish interface
Adds functions to request a filter by its enum ID, to open a file with a
preferred or the default filter, a function to get filtered samples and
another to close the chain.

Change %lld formats (which were wrong to begin with) to use PRIu64 from
inttypes.h.

Other minor fixes.