Get audio from AvxSynth as well.
Based on a patch by 0x09 (tab@lavabit.com). The original patch
description as follows:

In this case since they both target 2.5, the only modifications
to input/audio/avs.c are linking av[ix]synth instead of loading,
and not being usable as an importer for regular audio files (only
scripts)*.

*The other one allows things like

--audiofile somefile.mp3 --ademuxer avs

and opens somefile with (e.g.) DirectShowSource. Since all we have
is ffms2 it's kind of pointless.