Use liblsmash for L-SMASH-based mp4 support instead of the internal copy in output/mp4.
The internal copy has been removed.

The only caveat is that there's still a need for importer.h from
L-SMASH's source tree, and since this header is not installed with
L-SMASH, the user is required to manually copy it over themselves.
Since 'importer' is a bit of a generic name, the source files that
require it will be looking for lsmash_importer.h.

This could of course be avoided by patching L-SMASH to install its
headers to a subdirectory in the install path and put both lsmash.h
and importer.h there.  That's outside the scope of this patch, though.